feat: export mascot as GLB model and load via GLTFLoader
Replace programmatic Three.js geometry with a proper .glb model file
that can be opened in Blender. Added Node.js export script with browser
API polyfills for GLTFExporter, and rewrote the component to load the
model via GLTFLoader with named-part animations.
